skiping security manager test case in windows

+++ modules/integration/tests-integration/tests-security-manager/src/test/java/org/wso2/carbon/appserver/integration/test/server/security/manager/JavaSecurityManagerTestCase.java
@@ -18,6 +18,7 @@
 package org.wso2.carbon.appserver.integration.test.server.security.manager;
 
 import org.testng.Assert;
+import org.testng.SkipException;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
@@ -27,8 +28,10 @@ import org.wso2.appserver.integration.common.clients.WebAppAdminClient;
 import org.wso2.appserver.integration.common.utils.ASIntegrationTest;
 import org.wso2.appserver.integration.common.utils.WebAppDeploymentUtil;
 import org.wso2.appserver.integration.common.utils.WebAppTypes;
+import org.wso2.carbon.automation.engine.FrameworkConstants;
 import org.wso2.carbon.automation.engine.context.TestUserMode;
 import org.wso2.carbon.automation.engine.exceptions.AutomationFrameworkException;
+import org.wso2.carbon.automation.engine.frameworkutils.enums.OperatingSystems;
 import org.wso2.carbon.automation.test.utils.common.TestConfigurationProvider;
 import org.wso2.carbon.automation.test.utils.http.client.HttpRequestUtil;
 import org.wso2.carbon.automation.test.utils.http.client.HttpResponse;
@@ -58,6 +61,10 @@ public class JavaSecurityManagerTestCase extends ASIntegrationTest {
 
     @BeforeClass(alwaysRun = true)
     public void init() throws Exception {
+        if (System.getProperty(FrameworkConstants.SYSTEM_PROPERTY_OS_NAME)
+                .contains(OperatingSystems.WINDOWS.toString())) {
+            throw new SkipException("Skipping this test case in windows");
+        }
         super.init(userMode);
         webAppAdminClient = new WebAppAdminClient(backendURL, sessionCookie);
         webAppAdminClient.uploadWarFile(TestConfigurationProvider.getResourceLocation("AS")
@@ -175,6 +182,10 @@ public class JavaSecurityManagerTestCase extends ASIntegrationTest {
 
     @AfterClass(alwaysRun = true)
     public void clean() throws Exception {
+        if (System.getProperty(FrameworkConstants.SYSTEM_PROPERTY_OS_NAME)
+                .contains(OperatingSystems.WINDOWS.toString())) {
+            throw new SkipException("Skipping this test case in windows");
+        }
         webAppAdminClient.deleteWebAppFile(webAppFileName, hostName);
         //let webapp to undeploy
         Thread.sleep(2000);
